hex encode/decode pushed to ironfish-zkp (#5561)
